requirements.txt
requests
opencv-python
ultralytics
pandas
sqlalchemy
python-dotenv
jinja2
Pillow
apscheduler
matplotlib
psycopg2-binary
